# aspnetcore/src/IO.Swagger/Controllers/ExchangeControllerApi.cs
/*
 * Exchange API
 *
 * This service help to get exchange rate
 *
 * OpenAPI spec version: 2.0.1
 * Contact: manvendra80386@gmail.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ExchangeControllerApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <param name="from"></param>
        /// <param name="to"></param>
        /// <response code="200">default response</response>
        [HttpGet]
        [Route("/exchange/from/{from}/to/{to}")]
        [ValidateModelState]
        [SwaggerOperation("GetExchangerate")]
        [SwaggerResponse(statusCode: 200, type: typeof(ExchangeRate), description: "default response")]
        public virtual IActionResult GetExchangerate([FromRoute][Required]string from, [FromRoute][Required]string to)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ExchangeRate));
            string exampleJson = null;
            exampleJson = "{\n  \"exchangeId\" : \"exchangeId\",\n  \"rate\" : 0.8008281904610115,\n  \"from\" : \"from\",\n  \"to\" : \"to\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<ExchangeRate>(exampleJson)
                        : default(ExchangeRate);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}